requirements.txt
hdbcli
sqlalchemy
sqlalchemy-hana
python-dotenv
pandas
ai_core_sdk
flask